move hook to common util file
